Code cleanup for fix #4049
