+++ persons/models.py
@@ -3,6 +3,8 @@ from django.db import models
 from django.utils.translation import ugettext_lazy as _
 from django.core.exceptions import ValidationError
 from django.forms.fields import IntegerField
+from django.dispatch import receiver
+from django.db.models.signals import post_save
 
 from mks.models import Member, GENDER_CHOICES
 from .managers import PersonManager
@@ -91,6 +93,18 @@ class Person(models.Model):
             part.save()
         other.delete()
         self.save()
+
+
+@receiver(post_save, sender=Member)
+def member_post_save(sender, **kwargs):
+    instance = kwargs['instance']
+    person = Person.objects.get_or_create(mk=instance)[0]
+    for field in instance._meta.fields:
+        if field.name != 'id' and hasattr(person, field.name):
+            setattr(person, field.name, getattr(instance, field.name))
+
+    person.save()
+
 
 class Role(models.Model):
     text = models.CharField(blank=True,null=True, max_length=1024)
